requirements.txt
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
et-xmlfile==1.0.1
fire==0.2.1
graphviz==0.13
idna==2.8
imageio==2.5.0
ipdb==0.12.2
ipython==7.8.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.15.1
joblib==0.13.2
kiwisolver==1.1.0
matplotlib==3.1.1
networkx==2.3
numpy==1.17.1
openpyxl==2.6.3
pandas==0.25.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.3
pyzmq==18.1.0
requests==2.22.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
termcolor==1.1.0
torch==1.2.0
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.4.0
tornado==6.0.3
tqdm==4.35.0
traitlets==4.3.2
urllib3==1.25.3
visdom==0.1.8.8
wcwidth==0.1.7
websocket-client==0.56.0
